using post man for user data input and cloudinary

+++ src/utils/cloudinary.js
@@ -22,9 +22,14 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
+    console.log("CLOUDINARY ERROR:", error);
+
+    if (localFilePath && fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
+}
 }
